apply multipolygon ring decoding to MTV sources (previously only applied to polys generated by geojson-vt)
